fix breakage. bump version

+++ benchmark/carsales.rs
@@ -91,8 +91,8 @@ const MODELS : [&'static str; 6] = ["Camry", "Prius", "Volt", "Accord", "Leaf", 
 pub fn random_car(rng : &mut FastRand, mut car : car::Builder) {
     use std::mem::transmute;
 
-    car.set_make(MAKES[rng.next_less_than(MAKES.len() as u32) as uint]);
-    car.set_model(MODELS[rng.next_less_than(MODELS.len() as u32) as uint]);
+    car.set_make(MAKES[rng.next_less_than(MAKES.len() as u32) as usize]);
+    car.set_model(MODELS[rng.next_less_than(MODELS.len() as u32) as usize]);
 
     car.set_color(unsafe {transmute(rng.next_less_than(Color::Silver as u32 + 1) as u16) });
     car.set_seats(2 + rng.next_less_than(6) as u8);
